refactor(Views): move group creation to public router based on #174 discussions
