+++ api/views/__init__.py
@@ -1,7 +1,7 @@
 from api.views.age_range import AgeRangeViewSet
 from api.views.day_and_time_slot import DayAndTimeSlotViewSet
 from api.views.enrollment_test import EnrollmentTestResultViewSet, EnrollmentTestViewSet
-from api.views.group import GroupViewSet, PublicGroupViewSet
+from api.views.group import PublicGroupViewSet
 from api.views.language_and_level import LanguageAndLevelViewSet
 from api.views.non_teaching_help import NonTeachingHelpViewSet
 from api.views.personal_info import PersonalInfoViewSet
@@ -25,7 +25,6 @@ __all__ = [
     "PersonalInfoViewSet",
     "StudentViewSet",
     "PublicGroupViewSet",
-    "GroupViewSet",
     "PublicStudentViewSet",
     "PublicStudentWithPersonalInfoViewSet",
     "PublicTeacherViewSet",
