+++ api/views/group.py
@@ -4,6 +4,7 @@ from django.db import models
 from django_filters import rest_framework as filters
 from rest_framework import status, viewsets
 from rest_framework.decorators import action
+from rest_framework.mixins import CreateModelMixin
 from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.serializers import BaseSerializer
@@ -12,29 +13,13 @@ from api.filters import GroupFilter
 from api.models import Group
 from api.processors import GroupProcessor
 from api.serializers import (
-    GroupReadSerializer,
     GroupWriteSerializer,
     PublicGroupSerializer,
     PublicGroupWithStudentsSerializer,
 )
-from api.views.mixins import ReadWriteSerializersMixin
 
 
-class GroupViewSet(ReadWriteSerializersMixin, viewsets.ModelViewSet[Group]):  # type: ignore
-    """Viewset for groups. To be used by both coordinators in dashboard (public API) and by bot."""
-
-    queryset = Group.objects.all()
-    serializer_read_class = GroupReadSerializer
-    serializer_write_class = GroupWriteSerializer
-
-    def create(self, request: Request, *args: tuple[Any], **kwargs: dict[str, Any]) -> Response:
-        response = super().create(request, *args, **kwargs)
-        group = Group.objects.get(id=response.data["id"])
-        GroupProcessor.create(group)
-        return response
-
-
-class PublicGroupViewSet(viewsets.ReadOnlyModelViewSet[Group]):
+class PublicGroupViewSet(viewsets.ReadOnlyModelViewSet[Group], CreateModelMixin):
     """
     Public viewset for groups. Used for public API (Tooljet).
     """
@@ -48,6 +33,8 @@ class PublicGroupViewSet(viewsets.ReadOnlyModelViewSet[Group]):
             return PublicGroupSerializer
         if self.action == "retrieve":
             return PublicGroupWithStudentsSerializer
+        if self.action == "create":
+            return GroupWriteSerializer
 
         raise NotImplementedError(f"Unknown action: {self.action}")
 
@@ -63,3 +50,9 @@ class PublicGroupViewSet(viewsets.ReadOnlyModelViewSet[Group]):
         group = self.get_object()
         GroupProcessor.abort(group)
         return Response(status=status.HTTP_200_OK)
+
+    def create(self, request: Request, *args: tuple[Any], **kwargs: dict[str, Any]) -> Response:
+        response = super().create(request, *args, **kwargs)
+        group = Group.objects.get(id=response.data["id"])
+        GroupProcessor.create(group)
+        return response
